+++ packages/encode-lint-cli/.eslintrc.js
@@ -1,5 +1,5 @@
 module.exports = {
-  extends: ['null-eslint-config-encode/typescript/node', 'prettier'],
+  extends: ['rules-eslint-config/typescript/node', 'prettier'],
   rules: {
     '@typescript-eslint/no-require-imports': 0,
     'no-console': 0,