+++ packages/encode-lint-cli/src/lints/eslint/getESLintConfig.ts
@@ -25,6 +25,7 @@ export function getESLintConfig(opts: ScanOptions, pkg: PKG, config: Config): ES
   } else {
     // 根据扫描目录下有无lintrc文件，若无则使用默认的 lint 配置
     const lintConfigFiles = glob.sync('.eslintrc?(.@(js|yaml|yml|json))', { cwd })
+
     if (lintConfigFiles.length === 0 && !pkg.eslintConfig) {
       lintConfig.resolvePluginsRelativeTo = path.resolve(__dirname, '../../')
       lintConfig.useEslintrc = false